requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
timm>=0.9.12
albumentations>=1.4.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
einops>=0.7.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Vector Search (optional)
faiss-cpu>=1.7.2

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Development Tools (optional)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.4.0